Cleanup of the symmetry initialization of the phonon code.
Two main changes: The array irgq is removed.
                  The small group of q is calculated only in one routine.

The removal of the irgq array changed the list of input parameters of several
routines that may be called by external applications linked
with the phonon. I have tried not to break the compatibility
with these codes adding a suffix _new to the routines that required
a change of input parameters. The old routines are still there and
can be linked as before, but have been moved to a single file.
The old routines are:

smallgq -> becomes a routine called set_giq that sets only gi, gimq and
finds irotmq if minus_q is .true. It does not recalculate the small group
of q.

set_irr
set_irr_nosym
set irr_mode

are substituted by the two routines:

set_irr_new
set_irr_nosym_new

set_irr_mode is no more necessary. The input variables of the new routines
have been reduced. These routines set u, nirr, npert and these
are the only variables that are passed. All the others are passed by
USE association. Moreover these routines do not recalculate any more the
small group of q.

set_irr_sym -> set_irr_sym_new  (reduced the number of input parameters)

                               This routine sets t, tmq.                           

sgama_ph -> sgama_ph_new (the array sym is no more needed as input variable)

symdyn_munu -> symdyn_munu_new (the array irgq is no more in input variables)
symdynph_gq -> symdynph_gq_new (the array irgq is no more in input variables)
random_matrix -> random_matrix_new (the array irgq is no more in input variables)

dynmat0 -> dynmat0_new (call the new symdyn_munu_new)

dynmatrix -> dynmatrix_new (call the new symdyn_munu_new)

PAW_dusymmetrize (the array irgq has been removed from the list of input
variables, but the name of this routine has not been changed)

The old routines are in the file PHonon/PH/obsolete.f90. This file will
disapper in the future and the new routines will substitute the old ones,
so if you use these routines, you may update the argument list to the new
ones or copy the routines that you need in the directory of your
application (not advised) and change their names. Moreover soon the
array irgq will disappear from phcom.
Due to the changes in the organization of the symmetry routines in pw.x
now irgq(1)=1, irgq(2)=2, irgq(3)=3, ... and it is no more necessary to
use it. It can be safely removed from the routines that use it.
If you think that you need it, please create a copy in your common
variables. The array was set by smallgq.




git-svn-id: http://qeforge.qe-forge.org/svn/q-e/trunk/espresso@9273 c92efa57-630b-4861-b058-cf58834340f0