Fletcher-32 checksum implemented in clib/fletcher32.c
c-routine for a 16-byte integer array taken verbatim from the 
optimized c version reported in
https://en.wikipedia.org/wiki/Fletcher%27s_checksum#Optimizations
except for the fact that the data size is not an argument of the 
function but rather the address of its location is passed, as this
is what a fortran call passes as reference. 
The type of these variables is also defined as uint32_t 
instead of size_t because this is what fortran implicitely does.

Fortran interfaces to a generic subroutine fletcher32_cksum also
defined in UtilXlib/fletcher32_mod.f90



git-svn-id: http://qeforge.qe-forge.org/svn/q-e/trunk/espresso@13824 c92efa57-630b-4861-b058-cf58834340f0