WIth the new rho(G) file format, the Gamma-only case writes half G-vectors, the k-point case all G-vectors. Unfortunately this makes the rho(G) from the former case unreadable by the latter case. A single case is affected, I think: bands of a system whose scf is done with Gamma.
There is no easy workaround: the only solution is to store always rho(G) with
half G-vectors, Gamma or not Gamma, something that is convenient anyway, but
that may disrupt many pieces of code. For the time being, I just added a check
that yields an error with an understandable explanation of what happened why.

This might be the last change before 6.2 if no other bugs are found in the
next few hours.


git-svn-id: http://qeforge.qe-forge.org/svn/q-e/trunk/espresso@13914 c92efa57-630b-4861-b058-cf58834340f0