fix in case of lgamma ph computation with collected wave functions
in case the starting computation has been done twfcollect=.true.
run_nscf saves the wave functions in tmp_dir_phq and not in tmp_dir.
We have to find a way to have them in the same place in both cases.
Not now because release is tomorrow (29 june 2018).
Dirty fix if open_buffer fails in tmp_dir go back to tmp_dir_phq and try
again.