F90 flags passed to fox config in a cleaner way
exporting all QE F90 compilers flags may not work in many situations:

now a FOX_FLAGS is defined in make.inc, this is passed to FoX configure.
The content of this flag is set during configuration.
The configure script exports a @foxflags@ variable
which may be filled as needed. To program  the possible  content of the variable
one has to edit the m4/x_ac_qe_f90.m4  script.
As for now the only case when it is needed, is with the  PGFORTRAN compiler,
when the -Mlarge_arrays is set, in this for the FoX library to be
compatible it is necessary to compile also FoX with the same flag.