New implementation of the WS vectors in EPW
The new implementation relies on Wannier centers positions and is
triggered with the input variable "use_ws == .true."
When false, recover the old behavior.
No test has been added yet for this.

The update of the epw_polar test epw2 and epw3 is because of the
removal in EPW of the exponential look-up table which slightly change
numerical results.