IOLENGTH in EPW determined  runtime with INQUIRE
in epw some direct files are opened without using diropn, they were
using an hardcoded value for the IOLENGTH value. This was causing the
program to fail when the program was compiled for  KNL and SKL
machines.
As a temporary solution compute IOLENGTH with INQUIRE runtime.
More elegant solution use diropn to open files for direct access.