Calculate vdW kernel on the fly: first commit
Recent research such as PRB 99, 195418 (2019) opens the door for further
developments of the vdW-DF kernel and makes it desirable that the kernel
is calculated on the fly rather than read from file. Within the old
framework of using a kernel file, further improvements require to keep
track of different kernel files that all have the same name
(vdW_kernel_table), which becomes tedious and error prone.

In this first commit, I only focused on Modules/xc_vdW_DF.f90 and
incorporated the kernel calculation routines from
PW/src/generate_vdW_kernel_table.f90. Particular attention was given to
narrow the scope of the module subroutines and variables as much as
possible. In addition, I made other small coding adjustments and a
general cleanup of the entire module to adhere to the guidelines in the
QE developer manual---as a result, a "diff" of this file will show
substantial changes, even though scientifically the changes in the old
part of the code are minor.

At the state of this commit, the code computes the kernel in
Modules/xc_vdW_DF.f90, but only writes it to file so that we can compare
with the old (standard) kernel file. The calculation of the kernel is
numerically sensitive to the compilation details (see the note in the
header of GENERATE_KERNEL). As such, we find small differences in the
kernel file written from our new on-the-fly approach compared to the
standard qe-6.4.1 kernel file. The differences are typically a rounding
of the last digit, but can also be somewhat larger. In general, it seems
that differences appear at the 1E-12 level of kernel values or smaller.
However, these differences are at the same level as when comparing
kernels produced with the original PW/src/generate_vdW_kernel_table.f90
but with different numbers of processors. These kernel differences cause
only insignificant changes in Ecnl (typically less than 1E-10 Ryd) that
do not even show up in the output file (use verbosity = 'high' to see
the non-local correlation energy explicitly).

The kernel calculation even on a single core does not take very long
(~100 seconds on an Intel Xeon 2.20GHz, compiled with Intel 2018
compilers). However, the kernel calculation parallelizes perfectly and
on e.g. 8 cores takes only 12.5 seconds. I have added a line in
PW/src/print_clock_pw.f90 to print the time it takes to calculate the
vdW-DF kernel in the output file.

Note that in this commit, the kernel is calculated and written to file,
but not actually used in the calculation. As such, calculations will
show output identical to qe-6.4.1 calculations. In the next commit, we
will remove the writing to file part and use the calculated kernel in
the actual calculation.