Calculate vdW kernel on the fly: second commit
In this second commit I have now removed the reading of the kernel from
file and the kernel calculated on the fly is actually being used for the
vdW-DF calculation. Tests show that the output is identical to the
output produced by qe-6.4.1.

For the sake of compatibility with the old kernel file method and for
testing, I have included simple code at the bottom of generate_kernel to
write the on-the-fly calculated kernel to file or to alternatively read
a kernel from a kernel file (this code is commented out by default).

At this point, the xc_vdW_DF.f90 file is almost in its final form.
Small changes will occur in the next commit.