Calculate vdW kernel on the fly: fifth commit
This is the final commit for the vdW kernel on the fly recoding. The new
routines have been extensively tested and no differences were found when
compared to qe-6.4.1 output.

In this commit, we have updated the reference files for example16 of the
PHonon code. Finally, a short note describing this work was put into
Doc/release-notes.

This work was performed by T. Thonhauser and is supported by NSF Grant
No. 1712425.