Two problem (See below) related to the atomic-proj.xml were fixed (write_proj.f90). For fermi_proj.f90, unused variables are removed. Also, .frmsf files in the reference of the FermiSurfer example had incorrect Fermi velocity and reciprocal lattice vectors.
1, nspin=2 case
When atomic-proj.xml was written, the outermost loop was k and the second outermost was spin.
However, when it was read, the outermost loop was spin (the same as nks loop).
This file is also used in molecularpdos.f90, therfore I modified as in this commit.

2, nspin=4 case
For noncolinear case, read_atomic_proj routine raised 'mismatch' error because ns = 1 while ns_ = 4.