Added contribution by Hyungjun Lee
For full discussion and a working test/example,
see PR 295 on the Wannier90 repository:
https://github.com/wannier-developers/wannier90/pull/295

Additionally, I am adding to the gitignore one set of files
that is created by some compilers on Mac OS X.

Message from the original PR, by Hyungjun Lee:

This commit does the following three things:
(1) Fix the bug in the subroutine of compute_orb for the case with exclude_bands enabled.
(2) Modify the subroutine of compute_orb so as to avoid the unnecessary forward (inverse) FFTs.
(see #211)
(3) Additional minor modifications.

PS) Currently, compute_orb can't deal with the case with ultrasoft PPs and PAWs. However, even in the case using norm-conserving PPs, we have to maintain the storage for beta functions since non-local potential terms are calculated in h_psi.