vdW-DF: Implementation of an explicit h-function
The h-function inside the vdW-DF kernel provides an additional degree of
freedom for fine-tuning the functional performance of the vdW-DF
formalism---see e.g. PRB 99, 195418 (2019). To foster the implementation
of new functional developments within that family, it is useful to
define the h-function explicitly as a function call inside
Modules/xc_vdW_DF.f90. This work was performed by T. Thonhauser and is
supported by NSF Grant No. 1712425.