vdW-DF: adding test-suite examples and cleanup
We added test-suite examples for some of the new van der Waals
functionals. With the implementation of the kernel-on-the-fly vdW-DF
method, the individual tests run a little longer because the kernel is
being calculated every time. We thus cleaned-up the pw_vdw test-suite
examples so as to probe for fundamental functionality (testing for
different Z_ab values, stress, etc.) rather than testing all variants
individually. We also changed the naming convention to align with the
functional tested.