forces and stress printed only at  scf convergence
also other quantities needs to be checked for scf convergence
before saving in the xml data. But a thorough update of the
schema is needed.

Forces and stess updated now because there is no need to update
the schema for them and were causing the gpu version  to crash
when the program tried to print them before convergence.