Bug fix and cleaning of rigid.f90
There was a 2pi/a factor missing in the case of loto_2d = .true. missing.
This was found and explained in the preprint by M. Royo and M. Stengel [https://arxiv.org/abs/2012.07961].

Cleaning of the routine as well as addition of comments.
Removal of at(3,3) from the input variables.
Because intent(in) was not used, it was not detected that at(3,3) was not an input of the routine.
Also some reshuffling of terms to make them (hopefully) more clear.