More CUDA cleanup: g, gg
Module "gvect_gpum" deleted, replaced with copy of g, gg onto g_d, gg_d after
initialization of G-vectors, analogously to what is done with mill and mill_d.
NOTA BENE 1: it is not yet possible to move module recvec_subs into recvec and
make g_d and gg_d (and possibly other variables as well) "protected" variables,
because ggen is called, only once, in exx_bands, with a different set of
variables, and the code that does that is totally obscure :-(
NOTA BENE 2: I suspect that somewhere in CP G-vectors have to be rescaled for
variable-cell calculations, but I don't know where this is done. If not fixed,
this might break CP on GPU for variable-cell calculations.