Re-define (G+q) in long-range
From a physic point of view, (G+q) should be in unit of 1/Bohr.
It is much easier to deal with these units, especially when adding higher-order multipole.

I therefore change (G+q) from no unit to 1/Bohr unit by multiplying by 2pi/alat
Note that this is purely cosmetic for all cases except when quadrupole is use.
In such case there is a correct additional 2pi/alat factor.

In the case of rgd_blk_epw and rgd_blk_epw_fine the Ewald exponential should be adimentional.
However they are not at present because of this re-definition "qeq = qeq * twopi / alat".

For now I kept this inconsitency but we should make it adimensional in the future or (better) remove the
Ewald filter.