[CPV] bulk of the CG code ported to gpu.   NOT WORKING: restart from cg to verlet   NOT ACCELERATED: ultrasoft, ensemble dft
added useful debug/test module Modules/debug.f90
it works like the following: you put a call on the sub
CHECKPOINT(your_array/scalar_to_be_checked)

then, if in the input file, in the control namelist, the flag debug_checkpoints is .true., the routine does something.
If the flag debug_checkpoint_testing == .false., the data is written on the filesystem each time in a new file that is called
debug_checkpoint_file_prefix (input variables in the control namelist that defaults to 'debug_checkpoint'). Otherwise the data is read
from the filesystem and some useful information is displayed in the stdout, like the difference and the magnitude of the vectors.
This debug routine WORKS ONLY with a SINGLE MPI process