Enhance the robustness of `install/install_utils` script.
If the archive download fails due to force majeure, such as power and network failure, etc., the original script will refuse to run again due to the existence of a corrupted archive/package. So first check the integrity of the compressed file by gzip -t and delete it if damaged.