Oscure symmetrization problem
Under some unfortunate circumstances, the reciprocal-lattice symmetrization may fail. The reason is a combination of various factors:
- an atomic structure of questionable symmetry, typically with ibrav=0
- a questionable logic for G-vector ordering (see merge request !1773)

This patch works for the reported case, but depends upon the choice of a threshold that looks quite arbitrary to me. The ultimate solution is in my opinion to force exact symmetry upon the atomic structure