Debug for the dip-quad term of the IFC.
It only impacts the q=\Gamma and G\neq 0 part of the dynamical matrix
and on the tested cases, seems to have very little numerical impact.
It now follows Eqs. 46 and 47 of Phys. Rev. Research 3, 043022 (2021).
The EPW test-suite is unaffected and fully pass with this modification.
Bug found and reported by Jae-Mo Lihm.