ensemble dft tests.
checked that a molecular dynamics calculation produces the same trajectory with ensemble dft in cp.x and pw.x, for an 8 atom Al system, with gaussian smearing and both nlcc and ultrasoft pseudos. The difference between the trajectory was negligible in the tested range of ~0.5ps.

Added tests for ultrasoft and norm conserving.