Tentative implementation of PW's atomic wavefunction initialization inside CP
the original pw routine has the same interface, to not modify places where the old function is called.
The old code of the PW routine is moved inside a module that does not use PW global variables.
In the module I put also a wrapper for the cp code.

The atomic initialization in CP can save few iteration when the first minimization is done at the beginning of the simulation