Move the libxc module out of the dft_setting_routines, indirectly the XClib module, to functions.
When compiling C.f90 which uses module B which uses module A,
gfortran, ifort only need to see B.mod but nvfortran needs to see both A.mod and B.mod.
For this reason, if A and B belong to targets TA and TB respectively.
TB needs to use PUBLIC attribute when having a dependency on TA.
In this way, C.f90 finds the A.mod via the include path of A.mod propagated by TB.

If C.f90 doesn't need to know any internal detail about B.mod whether it uses A.mod or not,
it is better not to include A.mod in B.mod but just use A.mod in B.mod subroutines.
That is the method used in this patch.