change of timestep is performed automatically when needed
added dt of the last timestep (it can change with the autopilot module) in the xml file
the dt is read from the xml during the restart, and the code performs some additional
safety checks if the old method of changing the dt is used. If no change of timestep is
specified in the input, if the code detects that the old timestep and the new timesteps
differ more than 10^-6 hartree atomic units, the change of timestep is performed automatically
and a NOTE is printed in the output.

using of the old method is discouraged since it is error-prone

added test in the test suite to test a change of dt restart

removed documentation of the old method