removed duplicated code of the makov payne correction between CP and PW
added missing test of makov payne correction in CP
added missing test of makov payne correction for a molecule with dipole different from zero

the vacuum level calculation in CP was disabled (was not present before, and the number that it produces are strange)

new module makovpayne,
removed custom CP makovpayne code (note that in CP and PW tau has different usage. In PW it is the atomic coordinante in cell coordinate, and in CP in cartesian coordinates. In CP taus is the equivalent of tau in PW)

checked that CP and PW produces the same corrections for the ammonia input in the test suite