fix CP test suite, regenerating sio2-us-lda-3 and sio2-us-lda-4.in
commit b7845e8128cbd96c584517ce45d4b8a82847a025 introduced a new way of automatically handling the changing of the integration timestep on a CP restart.
the sio2-us-lda-2.in input have dt=15.0 and is a electron damping run, while
the sio2-us-lda-3.in input have dt=12.0 and should start with wfc velocities that are almost zero from the prev run

since the almost zero electronic velocities are rescaled (correctly) the initial condition for the lda-3 run are slightly different, producing slightly different results in the stress. The trajectories are with very good approximation almost equal.

The test suite was failing with the following:

cp_sio2 - sio2-us-lda-2.in: Passed.

cp_sio2 - sio2-us-lda-3.in: **FAILED**.
s1
    ERROR: absolute error 1.44e-03 greater than 5.00e-04. (Test: -5.25740981.  Benchmark: -5.25597336.)
    ERROR: relative error 2.73e-04 greater than 1.00e-04. (Test: -5.25740981.  Benchmark: -5.25597336.)

cp_sio2 - sio2-us-lda-4.in: **FAILED**.
s1
    ERROR: absolute error 1.40e-03 greater than 5.00e-04. (Test: 3.97451018.  Benchmark: 3.97590607.)
    ERROR: relative error 3.51e-04 greater than 1.00e-04. (Test: 3.97451018.  Benchmark: 3.97590607.)