Add the possibility to print el-ph matrix elements
Originally written by S. Ponc\'e and C. Verdi based on the elphsum
subroutine in 06/2018 and later modified by H. Lee in 06/2021.

Note that this commits _removes_ more lines than adding (422 lines added
and 1728 lines removed overall), thus improving readibility of the code.

In case the el-ph module is fully decoupled from PH in the future, it
should be moved with it but for now can be included there, I think.