Strange berry-phase problem with empty bands
Incorrect (but irrelevant) diagonal matrix elements could lead to overflows
and NaN's in Berry phase and similar for many empty bands and processors.
NOTE 1: Berry-phase and similar calculations deal with empty bands in a funny
way: an array (l_cal) keeps track of which bands are empty so that matrix
elements between empty bands are not computed and set to zero (off-diagonal)
or one (diagonal). The determinant of the (nbnd*nbnd) matrix is then computed.
It would be easier and faster to compute nbnd_occ, number of occupied bands,
and to perform the calculation for the (nbnd_occ*nbnd_occ) matrix only.
NOTE 2: the determinant is computed in several places, using two different ways:
linpack zgefa + zgedi or lapack zgetrf