Last DEVICE variable "qq_nt_d" in upflib removed
This variable, used only in CP, "ortho" routine, is replaced by openACC.
There was an obscure section of code that I haven't ported, though: if qq_nt_d
was smaller than 10^-5, it was set to zero. Something similar is present in the
non-GPU code: if qq_nt is smaller than 10^-5, it is not set to zero but its
contribution to a term of "ortho" is not computed. I don't know wy this is done
and how important it is. I don't think it is that important, though. It might
have been a trick to soare a few CPU cycles.

Another obscure point is the following: there are two sets of similar variables,
qq_nt and qq_at, the former depending on type of atoms, the latter on atoms.
The latter are the same as the former: having both is confusing and misleading.