Simpson integration with even number of points
The original routine, devised to work with an odd number of grid points,
contained a workaround for the case of even number of grid points that did
not work (around). See issue #802. Usually not a big deal because just the
final point was neglected, and because PP files usually were forced to have
an odd number of grid points, but in some cases the effect was not negligible.
This version implements the same routine as in DFTK (see issue #802).
Not sure how good it is, but it is for sure better than the previous one.